requirements.txt
# STF-Net Requirements
# Tested with Python 3.10+ on CUDA 11.8 and CUDA 12.1
# Install with: pip install -r requirements.txt

# Core deep learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0   # optional, included for environment compatibility

# Backbone and pretrained models
timm>=0.9.12

# Face detection (MTCNN)
facenet-pytorch>=2.5.3

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0   # for SSIM computation

# Data and numerics
numpy>=1.24.0
scipy>=1.11.0

# Evaluation metrics
scikit-learn>=1.3.0

# Configuration
PyYAML>=6.0.1

# Progress bars and logging
tqdm>=4.66.0

# Visualisation (optional, for Grad-CAM overlays)
matplotlib>=3.7.0
